refactor(tui): split render helpers by concern
